requirements.txt
Django==5.0.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.6
django-celery-beat==2.6.0
requests==2.31.0
drf-spectacular==0.27.1
python-decouple==3.8
dj-database-url==1.2.0
